requirements.txt
alembic==0.9.6
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.2
Babel==2.7.0
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.8.1
billiard==3.3.0.23
bleach==2.1.3
blinker==1.4
blis==0.7.4
boto==2.49.0
boto3==1.10.15
botocore==1.13.15
bz2file==0.98
catalogue==2.0.1
celery==3.1.23
certifi==2019.9.11
cffi==1.14.5
chardet==3.0.4
cheroot==8.2.1
CherryPy==18.4.0
click==7.1.2
colorama==0.4.1
coverage==4.1
cryptography==3.2.1
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.2
cytoolz==0.10.1
decorator==4.4.1
docutils==0.15.2
dominate==2.3.1
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==0.9.1
future==0.18.2
gensim==3.8.1
gnureadline==8.0.0
greenlet==1.0.0
gunicorn==19.7.1
html5lib==1.1
httplib2==0.10.3
idna==2.8
ImageHash==3.7
importlib-metadata==3.7.3
inca==0.1
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jaraco.functools==2.0
jedi==0.15.1
jieba3k==0.35.1
Jinja2==2.11.3
jmespath==0.9.4
joblib==0.14.0
jsonschema==2.6.0
jwt==1.1.0
kiwisolver==1.1.0
kombu==3.0.37
loremipsum==1.0.5
lxml==4.4.1
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==0.23
matplotlib==3.1.1
more-itertools==7.2.0
murmurhash==1.0.2
mysqlclient==1.4.5
networkx==2.4
newspaper3k==0.2.8
nl-core-news-sm==2.3.0
nltk==3.2.4
numpy==1.17.4
oauth2client==4.1.2
oauthlib==1.1.2
packaging==20.9
pandas==0.25.3
parso==0.5.1
pathy==0.4.0
patsy==0.4.1
Pattern==3.6
pdfminer.six==20191110
pexpect==4.7.0
pickleshare==0.7.2
Pillow==6.2.0
plac==0.9.6
portend==2.6
praw==7.1.0
prawcore==1.5.0
preshed==3.0.5
prompt-toolkit==2.0.10
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.20
pycryptodome==3.9.3
pydantic==1.7.3
Pygments==2.4.2
pyparsing==2.4.5
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==0.6.2
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.2
selenium==3.141.0
six==1.13.0
smart-open==3.0.0
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==3.0.5
spacy-legacy==3.0.1
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
srsly==2.4.0
statsmodels==0.10.1
tempora==1.14.1
thinc==8.0.2
tinysegmenter==0.3
tldextract==2.2.2
toolz==0.10.0
tornado==4.2
tqdm==4.38.0
traitlets==4.3.3
twython==3.4.0
typer==0.3.2
typing-extensions==3.7.4.3
ua-parser==0.10.0
update-checker==0.18.0
urllib3==1.24.3
user-agents==2.2.0
vine==1.3.0
visitor==0.1.3
wasabi==0.8.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
WTForms==2.1
zc.lockfile==2.0
zipp==3.4.1
